+++ Project/Engine/CRenderMgr.cpp
@@ -115,29 +115,29 @@ void CRenderMgr::UpdateData()
 {
 
     // 구조화버퍼의 크기가 모자라면 더 크게 새로 만든다.
-    if (m_Light2DBuffer->GetElementCount() < m_vecLight2DInfo.size())
+    if (m_Light2DBuffer->GetElementCount() < (UINT)m_vecLight2DInfo.size())
     {
-        m_Light2DBuffer->Create(sizeof(tLightInfo), m_vecLight2DInfo.size(), SB_TYPE::READ_ONLY, true);
+        m_Light2DBuffer->Create(sizeof(tLightInfo), (UINT)m_vecLight2DInfo.size(), SB_TYPE::READ_ONLY, true);
     }
 
     // 구조화버퍼로 광원 데이터를 옮긴다.
-    m_Light2DBuffer->SetData(m_vecLight2DInfo.data(), sizeof(tLightInfo) * m_vecLight2DInfo.size());
+    m_Light2DBuffer->SetData(m_vecLight2DInfo.data(), sizeof(tLightInfo) * (UINT)m_vecLight2DInfo.size());
 
     m_Light2DBuffer->UpdateData(12, PIPELINE_STAGE::PS_PIXEL);
 
     if (m_Light3DBuffer->GetElementCount() < m_vecLight3DInfo.size())
     {
-        m_Light3DBuffer->Create(sizeof(tLightInfo), m_vecLight3DInfo.size(), SB_TYPE::READ_ONLY, true);
+        m_Light3DBuffer->Create(sizeof(tLightInfo), (UINT)m_vecLight3DInfo.size(), SB_TYPE::READ_ONLY, true);
     }
 
     // 구조화버퍼로 광원 데이터를 옮긴다.
-    m_Light3DBuffer->SetData(m_vecLight3DInfo.data(), sizeof(tLightInfo) * m_vecLight3DInfo.size());
+    m_Light3DBuffer->SetData(m_vecLight3DInfo.data(), sizeof(tLightInfo) * (UINT)m_vecLight3DInfo.size());
     m_Light3DBuffer->UpdateData(13, PIPELINE_STAGE::PS_PIXEL);
 
 
     // GlobalData 에 광원 개수정보 세팅
-    GlobalData.Light2DCount = m_vecLight2D.size();
-    GlobalData.Light3DCount = m_vecLight3D.size();
+    GlobalData.Light2DCount = (UINT)m_vecLight2D.size();
+    GlobalData.Light3DCount = (UINT)m_vecLight3D.size();
 
     // 전역 상수 데이터 바인딩
     CConstBuffer* pGlobalBuffer = CDevice::GetInst()->GetConstBuffer(CB_TYPE::GLOBAL);
