+++ Project/Engine/CRenderMgr.h
@@ -48,14 +48,14 @@ public:
     { 
         m_vecLight2D.push_back(_Light2D); 
         m_vecLight2DInfo.push_back(_Info);  
-        return m_vecLight2D.size() - 1;
+        return (UINT)m_vecLight2D.size() - 1;
     }
 
     UINT RegisterLight3D(CLight3D* _Light3D, const tLightInfo& _Info) 
     { 
         m_vecLight3D.push_back(_Light3D); 
         m_vecLight3DInfo.push_back(_Info); 
-        return m_vecLight3D.size() - 1; 
+        return (UINT)m_vecLight3D.size() - 1;
     }
 
     void ClearCamera() { m_vecCam.clear(); }
